Reintroduced homepage preference on GSettings
This setting is particulary important for the kiosk mode
of Epiphany where it's neccesary that the browser is opened
starting on a very specific home address.

There is not a need to create a UI for the configuration of the
setting as on the mentioned main purposes of the flag like
the mentioned kiosk mode, or intranets, the homepage-url, would
be basically system-configured.

https://bugzilla.gnome.org/show_bug.cgi?id=702435