Display documents inside ephy window using libevince
This patch ports Carlos Garcia Campos' patch in order to support it in
gnome-3-8 branch.

As Carlos mentions, it adds a new widget EphyDocumentView to load
documents using EvView. When the policy checker detects the document
can't be shown inside a web view, and it's a document supported by
evince, a download is started for the document. The EphyEmbed mode is
set to DOCUMENT so that when the download finishes, a EphyDocumentView
is created to load the downloaded document.

In EphyWindow we consider documents as about:blank pages for now, so
that all actions (search, copy, paste, etc.) are disabled.

https://bugzilla.gnome.org/show_bug.cgi?id=689992