Removed undesired shortcuts and included a popup to show the available
The undesired shorctuts that were not excluded by the lockdown settings
were manually removed, and a new tooltip is shown when pressing F1
to populate with the list of supported keys.
Currenty is using the same GTK_LABEL that is being used by the fullscreen
command. It migh be interesting to use a different variable for
the shortcuts messages.