Hide location bar when homepage-url is not set
In epy location-bar is part of the toolbar abstraction. This patch complements
the patch "Hide toolbars when 'disable-menus' (8ca3330904d3e9b827c0608403f04f)

location-bar gets always painted under a GtkBox. This GtkBox is coupled with
another abstractions, as location controller/signals/etc, so the best approach
looks keeping it but not painting it instead of adapting all code depending on
it.

To achieve the previous approach I am packing it when disable-menus is false
only. This approach is coherent with the original patch mentioned above.